feat: update point float64

+++ internal/pkg/dal/rao/scene.go
@@ -57,8 +57,8 @@ type SaveFlowResp struct {
 }
 
 type Point struct {
-	X int `json:"x"`
-	Y int `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 
 type Node struct {
